Not CDNs
CDNs operate in a number of different ways, and the pattern that's being described here (site maintains control of HTML and offloads embedded content to CDN) is a very small subset of that. 

Using the term like this is going to be confusing for folks who use CDNs in a different way, and the underlying concern is really third-party content (as you say, when "it's delivered from a server outside their control.")

CDNs don't generally operate in the way that's being described here.