Drop the `alg` parameter.
Section 7.3.6 of RFC9421 suggests that the alg parameter be included only
if it's a relevant join in the application's logic, discouraging its
usage in general ("... applications are encouraged to use other mechanisms
such as static configuration or a higher-protocol-level algorithm
specification instead, preventing an attacker from substituting the
algorithm specified."). Given that we only support Ed25519, it probably
makes sense to drop the parameter from our profile.

Closes WICG/signature-based-sri#33.